resume_extractor/requirements.txt
fastapi
uvicorn
streamlit
pdfplumber
spacy
pandas
mysql-connector-python
matplotlib
altair==4.2.2